load jquery differently; remove the fill_frame module because it wasnt really modular; use rollup/babel/es6; spelling
