+++ index-es6.html
@@ -80,7 +80,8 @@
 				    </tr>
 				</table>
 			</div>
-			
+
+			<script src="https://code.jquery.com/jquery-2.2.0.min.js"></script>
 			<script src="dist-final.js"></script>			
 		</div>
 		<hr></hr>
