+++ js-es6/jscolor-1.4.3/jscolor.js
@@ -10,9 +10,7 @@
  */
 
 
-export default {
-
-
+let jscolor = {
 	dir : '', // location of jscolor directory (leave empty to autodetect)
 	bindClass : 'color', // class name
 	binding : true, // automatic binding via <input class="...">
@@ -990,5 +988,7 @@ export default {
 
 		this.importColor();
 	}
-
 };
+
+export default jscolor;
+
